+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Selected elements update modes
+    - Removed `static` access to `\Mds\PimPrint\CoreBundle\Service\ProjectsManager::isLocalizedProject`
+
 ## [3.2.0] - 2023-11-27
 
 ### Feature